M11: Items 50-57 — inventory/trade/quest/shop/group protocol fixes
Item 50: Trade item-duplication guard; mail equipped-item bypass
Item 51: Movement broadcast flood; teleport jump animation; post-teleport spawn sequence
Item 52: Air-movement flood; cross-zone revive spawn; loot-close broadcast; respawn home position
Item 53: Dead-player consumable exploit; SM_LOOT_STATUS enum completeness
Item 54: Regen broadcast; GM heal broadcast; GM cross-zone tp appearance
Item 55: Item.Slot semantics fix (explicit IsEquipped bool); CM_DELETE_ITEM equip guard;
         CM_CRAFT partial notification; CM_SPLIT_ITEM full-stack split; V10 migration
Item 56: SM_INVENTORY_INFO equip filter; CM_BUY_ITEM kinah error message;
         FindByItemId equip guard; quest selectable reward items (extendedRewardIndex)
Item 57: SM_GROUP_INFO wire format matched to Java (group metadata only);
         SM_GROUP_MEMBER_INFO rewritten with GroupEvent enum (Join/Enter/Update/Leave/
         Disconnected) and DWORD HP/MP, flight points, class, gender, event tail;
         CM_INVITE_TO_GROUP sends correct JOIN + ENTER sequence; GsClientConnection
         disconnect uses SM_GROUP_MEMBER_INFO(Disconnected) instead of SM_GROUP_INFO

+++ AionLightning.NET/AionLightning.Chat/Network/Gs/GsConnection.cs
@@ -50,7 +50,7 @@ public sealed class GsConnection : AConnection
         int wireLen = 2 + 1 + bodyBuf.WrittenCount;
         byte[] wire = new byte[wireLen];
         BinaryPrimitives.WriteInt16LittleEndian(wire, (short)wireLen);
-        wire[2] = packet.Opcode;
+        wire[2] = (byte)packet.Opcode;
         bodyBuf.WrittenSpan.CopyTo(wire.AsSpan(3));
 
         await WriteRawAsync(wire, ct);
