+++ AionLightning.NET/AionLightning.Game/Network/Ls/LsConnection.cs
@@ -62,7 +62,7 @@ public sealed class LsConnection : AConnection
 
         byte[] wire = new byte[wireLen];
         BinaryPrimitives.WriteInt16LittleEndian(wire, (short)wireLen);
-        wire[2] = packet.Opcode;
+        wire[2] = (byte)packet.Opcode;
         bodyBuf.WrittenSpan.CopyTo(wire.AsSpan(3));
 
         await WriteRawAsync(wire, ct);
